# Faster bitmask iteration