--- a/arrow/src/util/bit_chunk_iterator.rs
+++ b/arrow/src/util/bit_chunk_iterator.rs
@@ -151,7 +151,7 @@ impl<'a> UnalignedBitChunk<'a> {
         self.chunks
     }
 
-    pub fn iter(&self) -> UnalignedBitChunkIterator<'a> {
+    pub(crate) fn iter(&self) -> UnalignedBitChunkIterator<'a> {
         self.prefix
             .into_iter()
             .chain(self.chunks.iter().cloned())
@@ -164,7 +164,7 @@ impl<'a> UnalignedBitChunk<'a> {
     }
 }
 
-pub type UnalignedBitChunkIterator<'a> = std::iter::Chain<
+pub(crate) type UnalignedBitChunkIterator<'a> = std::iter::Chain<
     std::iter::Chain<
         std::option::IntoIter<u64>,
         std::iter::Cloned<std::slice::Iter<'a, u64>>,