# Which issue does this PR close?

Closes #1227.

# Rationale for this change
 
This improves the filter benchmarks by a factor of 2x, and likely will have similar benefits elsewhere

```
filter u8               time:   [140.44 us 140.61 us 140.76 us]                      
                        change: [-51.558% -51.392% -51.226%] (p = 0.00 < 0.05)
                        Performance has improved.

filter u8 high selectivity                                                                             
                        time:   [2.4576 us 2.4587 us 2.4601 us]
                        change: [-53.091% -52.977% -52.863%] (p = 0.00 < 0.05)
                        Performance has improved.

filter u8 low selectivity                                                                             
                        time:   [1.6956 us 1.6981 us 1.7010 us]
                        change: [-60.543% -60.284% -59.829%] (p = 0.00 < 0.05)
                        Performance has improved.
                        
filter f32              time:   [418.84 us 419.45 us 420.15 us]                       
                        change: [-26.414% -26.286% -26.141%] (p = 0.00 < 0.05)
                        Performance has improved.

filter single record batch                                                                            
                        time:   [183.96 us 188.44 us 193.16 us]
                        change: [-33.234% -31.549% -29.597%] (p = 0.00 < 0.05)
                        Performance has improved.

```
Note: the filter **context** benchmarks don't see any change as the filter is computed outside the benchmark body.

# What changes are included in this PR?

This adds an UnalignedBitChunkIterator and updates SlicesIterator to use it

# Are there any user-facing changes?

No