@tustvold 2022-01-28
I've rebased this to not be on top of #1225 as it is pending some more thought, without that fix this makes a smaller delta but still not insignificant

```
filter u8               time:   [359.76 us 359.84 us 359.93 us]                      
                        change: [-27.199% -27.161% -27.119%] (p = 0.00 < 0.05)
                        Performance has improved.

filter u8 high selectivity                                                                             
                        time:   [8.8616 us 8.9026 us 8.9439 us]
                        change: [-29.670% -29.463% -29.275%] (p = 0.00 < 0.05)
                        Performance has improved.

filter u8 low selectivity                                                                             
                        time:   [2.4243 us 2.4249 us 2.4254 us]
                        change: [-45.812% -45.627% -45.478%] (p = 0.00 < 0.05)
                        Performance has improved.
```

Perhaps more importantly I hope to use this to implement optimized filter kernels