@tustvold 2022-01-29
I think SlicesIterator has a bug :disappointed: - fixing...